+++ rest-api-spec/src/yamlRestTest/resources/rest-api-spec/test/tsdb/15_timestamp_mapping.yml
@@ -164,7 +164,7 @@ reject timestamp meta field with wrong type:
       reason: introduced in 8.0.0 to be backported to 7.16.0
 
   - do:
-      catch: /.* time series index \[_data_stream_timestamp\] meta field must be enabled/
+      catch: /\[_data_stream_timestamp\] meta field has been disabled/
       indices.create:
         index: test
         body:
